Angular 12 upgrade
This PR upgrades yti-common-ui to be compatible with Angular 12. The
angular upgrade was originally done to fix small issues, but has since
grown into a larger effort.

The final commit is squashed from several separate commits to make it
clear when a change originates from this upgrade.

Updated dependencies:

* Primary upgrade has been Angular to version 12.
* Another necessary upgrade was ng-bootstrap to version 10.0.0, and Bootstrap to 4.5.0
* Other library versions updated to reasonably recent versions

Library project:

The project was previously included as-is, but is now converted to an
Angular Library project.

https://angular.io/guide/creating-libraries

This means the main source files have been moved to project/, and some
basic project configuration like angular.json have been recreated
using the latest angular defaults.

Notification component:

The notification component uses some internal code from bootstrap for
popups, and was causing some problems. Since the component only seems to
be used by yti-groupmanagement, it has been removed from the project
entirely, at least for now.

require():

Usages of require() have been replaced with import or with other means.

Packaging changes

The angular upgrade, along with adoption of Ivy, requires that the
library is fully compiled before usage.

Previously the package.json could include the library like this:

  "yti-common-ui": "https://github.com/VRK-YTI/yti-common-ui.git/#22f82191de35575bd571b94ce0447cbe54faae43",

Now this will not work, and we must include the compiled output instead.

Either include the output directory:

  "@vrk-yti/yti-common-ui": "file:///home/example/projects/dvv/yti/yti-common-ui/dist/yti-common-ui/"

Or npm package, published in the github registry:

  "@vrk-yti/yti-common-ui": "0.0.1-alpha-20210906.g533e324.feature-angular-library",

The npm package source for the @vrk-yti scope is configured in .npmrc
and .yarnrc, so it should never attempt to load from public registry.

publish.js can be used to publish a new release to the github npm
registry. See README.md for details

https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-npm-registry

Version numbering:

As part of publishing the npm package, we must have a meaningful version number.

publish.js now contains the logic for generating a version number, which
looks different depending on the branch (master/develop/other):

* 0.0.1
* 0.0.1-beta+20210906.ebaf822d
* 0.0.1-alpha+20210906.ebaf822d.wip-branch